collision still dosn't work, but you can't go too fast.

+++ Unnamed Game/Platform.h
@@ -1,9 +1,5 @@
 #pragma once
 #include"Globals.h"
-#include<allegro5/allegro.h>
-#include<allegro5/allegro_image.h>
-#include<allegro5/allegro_primitives.h>
-#include<cmath>
 class Platform{
 private:
 	double Xpos;
@@ -14,7 +10,7 @@ private:
 	ALLEGRO_BITMAP* pic;
 public:
 	Platform(int X, int Y, int Breakable);
-	int collision(int X, int Y, bool Distance);
+	int collision(double X, double Y, bool Distance);
 	void move();
 	void draw();
 };
